[REFACTOR] enum인 Category 클래스 위치 변경
